refactor: move `PageRangeExt` out of `mm`
