+++ src/mm/physicalmem.rs
@@ -4,8 +4,6 @@ use core::sync::atomic::{AtomicUsize, Ordering};
 
 #[cfg(feature = "hermit-entry")]
 use align_address::Align;
-#[cfg(feature = "hermit-entry")]
-use free_list::PageRangeError;
 use free_list::{FreeList, PageLayout, PageRange};
 use hermit_sync::InterruptTicketMutex;
 use memory_addresses::VirtAddr;
@@ -19,6 +17,8 @@ use crate::arch::mm::paging::{self, HugePageSize, PageSize};
 use crate::env::{self, FdtStartInfo};
 use crate::mm::device_alloc::DeviceAlloc;
 use crate::mm::{PageRangeAllocator, PageRangeBox};
+#[cfg(feature = "hermit-entry")]
+use crate::page_range_ext::PageRangeExt;
 
 static PHYSICAL_FREE_LIST: InterruptTicketMutex<FreeList<16>> =
 	InterruptTicketMutex::new(FreeList::new());
@@ -188,29 +188,6 @@ unsafe fn detect_from_fdt() -> Result<(), ()> {
 	reserve(fdt_region);
 
 	Ok(())
-}
-
-// FIXME: upstream these
-#[cfg(feature = "hermit-entry")]
-trait PageRangeExt: Sized {
-	fn containing(start: usize, end: usize) -> Result<Self, PageRangeError>;
-
-	fn and(self, rhs: Self) -> Option<Self>;
-}
-
-#[cfg(feature = "hermit-entry")]
-impl PageRangeExt for PageRange {
-	fn containing(start: usize, end: usize) -> Result<Self, PageRangeError> {
-		let start = start.align_down(free_list::PAGE_SIZE);
-		let end = end.align_up(free_list::PAGE_SIZE);
-		Self::new(start, end)
-	}
-
-	fn and(self, rhs: Self) -> Option<Self> {
-		let start = self.start().max(rhs.start());
-		let end = self.end().min(rhs.end());
-		Self::new(start, end).ok()
-	}
 }
 
 unsafe fn init() {
